# docs/api/openapi/storefront.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/storefront": {
      "get": {
        "operationId": "StorefrontController_getStorefront",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicStorefrontDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Get public storefront data",
        "tags": [
          "Storefront"
        ]
      }
    },
    "/storefront/products": {
      "get": {
        "operationId": "StorefrontController_listProducts",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by keyword",
            "schema": {
              "example": "MacBook",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "default": "WEBSITE",
              "type": "string",
              "enum": [
                "POS",
                "WEBSITE",
                "FACEBOOK",
                "INSTAGRAM",
                "TIKTOK"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "example": "coffee",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicProductDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "List publicly visible products",
        "tags": [
          "Storefront"
        ]
      }
    },
    "/storefront/products/{productSlug}": {
      "get": {
        "operationId": "StorefrontController_getProduct",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "productSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by keyword",
            "schema": {
              "example": "MacBook",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "default": "WEBSITE",
              "type": "string",
              "enum": [
                "POS",
                "WEBSITE",
                "FACEBOOK",
                "INSTAGRAM",
                "TIKTOK"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "example": "coffee",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProductDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Get a publicly visible product",
        "tags": [
          "Storefront"
        ]
      }
    },
    "/storefront/theme": {
      "get": {
        "operationId": "StorefrontController_getTheme",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveThemeDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Get the live storefront theme",
        "tags": [
          "Storefront"
        ]
      }
    },
    "/storefront/orders": {
      "get": {
        "operationId": "StorefrontController_listOrders",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by keyword",
            "schema": {
              "example": "MacBook",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "customerEmail",
            "required": true,
            "in": "query",
            "schema": {
              "example": "jane@example.com",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_PAYMENT",
                "RESERVED",
                "PAID",
                "PROCESSING",
                "FULFILLED",
                "COMPLETED",
                "CANCELLED",
                "PAYMENT_FAILED",
                "EXPIRED",
                "REFUNDED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "List a customer's orders by email, optionally filtered by status",
        "tags": [
          "Storefront"
        ]
      }
    },
    "/storefront/orders/{orderNumber}": {
      "get": {
        "operationId": "StorefrontController_getOrder",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "required": true,
            "in": "query",
            "schema": {
              "example": "jane@example.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Get order detail by order number, verified against the customer email",
        "tags": [
          "Storefront"
        ]
      }
    },
...
    "/storefront/orders/{orderNumber}/tracking": {
      "get": {
        "operationId": "StorefrontDeliveryController_track",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "required": true,
            "in": "query",
            "description": "Must match the email the order was placed with",
            "schema": {
              "format": "email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Track an order, verified against the customer email",
        "tags": [
          "Delivery"
        ]
      }
    },
...
    "/social-links/{hotspotId}": {
      "get": {
        "operationId": "PublicSocialController_resolveSocialLink",
        "parameters": [
          {
            "name": "hotspotId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "schema": {
              "default": "WEBSITE",
              "type": "string",
              "enum": [
                "WEBSITE",
                "FACEBOOK",
                "INSTAGRAM",
                "TIKTOK"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialLinkDto"
                }
              }
            }
          }
        },
        "security": [
          {}
        ],
        "summary": "Resolve a hotspot with current stock availability",
        "tags": [
          "Social Commerce"
        ]
      }
    },
    "/storefront/posts": {
      "get": {
        "operationId": "PublicSocialController_listArticles",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "List published website articles",
        "tags": [
          "Social Commerce"
        ]
      }
    },
    "/storefront/posts/{slug}": {
      "get": {
        "operationId": "PublicSocialController_getArticle",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Get a published website article",
        "tags": [
          "Social Commerce"
        ]
      }
    }
  },
  "info": {
    "title": "Storefront API",
    "description": "Storefront browsing, carts, shopper addresses (guest via cart token, or signed-in via bearer token), delivery options and tracking, checkout sessions, payment initiation, and payment status polling",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "checkout-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Checkout-Token"
      },
      "cart-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cart-Token"
      }
    },
    "schemas": {
...
      "PublicMerchantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "LiveThemeDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          },
          "publishedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "version",
          "config"
        ]
      },
      "PublicProductCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "sortOrder"
        ]
      },
      "PublicProductMediaDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "url",
          "type",
          "sortOrder"
        ]
      },
      "PublicProductVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "29.99"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "isAvailable": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "sku",
          "price",
          "attributes",
          "isAvailable"
        ]
      },
      "PublicProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "sku": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "29.99"
          },
          "currency": {
            "type": "string"
          },
          "channel": {
            "type": "string",
            "enum": [
              "POS",
              "WEBSITE",
              "FACEBOOK",
              "INSTAGRAM",
              "TIKTOK"
            ]
          },
          "categoryId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "category": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicProductCategoryDto"
              }
            ]
          },
          "baseIsAvailable": {
            "type": "boolean"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "isPurchasable": {
            "type": "boolean"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicProductMediaDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicProductVariantDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "sku",
          "price",
          "currency",
          "channel",
          "baseIsAvailable",
          "isAvailable",
          "isPurchasable",
          "media",
          "variants"
        ]
      },
      "PublicStorefrontDto": {
        "type": "object",
        "properties": {
          "merchant": {
            "$ref": "#/components/schemas/PublicMerchantDto"
          },
          "theme": {
            "$ref": "#/components/schemas/LiveThemeDto"
          },
          "featuredProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicProductDto"
            }
          }
        },
        "required": [
          "merchant",
          "theme",
          "featuredProducts"
        ]
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "object",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "string"
          },
          "totalPrice": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "sku",
          "name",
          "quantity",
          "unitPrice",
          "totalPrice"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "sourceChannel": {
            "type": "string",
            "enum": [
              "POS",
              "WEBSITE",
              "FACEBOOK",
              "INSTAGRAM",
              "TIKTOK"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_PAYMENT",
              "RESERVED",
              "PAID",
              "PROCESSING",
              "FULFILLED",
              "COMPLETED",
              "CANCELLED",
              "PAYMENT_FAILED",
              "EXPIRED",
              "REFUNDED"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "PARTIAL",
              "PAID",
              "FAILED",
              "REFUNDED"
            ]
          },
          "fulfillmentStatus": {
            "type": "string",
            "enum": [
              "UNFULFILLED",
              "PROCESSING",
              "FULFILLED",
              "CANCELLED"
            ]
          },
          "subtotalAmount": {
            "type": "string"
          },
          "discountAmount": {
            "type": "string"
          },
          "feeAmount": {
            "type": "string"
          },
          "totalAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "id",
          "orderNumber",
          "sourceChannel",
          "status",
          "paymentStatus",
          "fulfillmentStatus",
          "subtotalAmount",
          "discountAmount",
          "feeAmount",
          "totalAmount",
          "currency",
          "items"
        ]
      },
...
      "SocialLinkDto": {
        "type": "object",
        "properties": {
          "hotspotId": {
            "type": "string",
            "format": "uuid"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "product": {
            "type": "object"
          },
          "storefrontPath": {
            "type": "string"
          }
        },
        "required": [
          "hotspotId",
          "isAvailable",
          "product",
          "storefrontPath"
        ]
      }
    }
  }
}